refactor: made changes to routing
